+++ src/index.ts
@@ -4,6 +4,8 @@ export * from './PropertyFieldListPicker';
 export * from './PropertyFieldPeoplePicker';
 export * from './PropertyFieldSpinButton';
 export * from './PropertyFieldTermPicker';
+export * from './PropertyFieldMultiSelect';
+export * from './PropertyFieldNumber';
 
 export * from './PropertyFieldButtonWithCallout';
 export * from './PropertyFieldCheckboxWithCallout';