+++ src/propertyFields/multiSelect/IPropertyFieldMultiSelectHost.ts
@@ -1,0 +1,7 @@
+import { IDropdownProps } from 'office-ui-fabric-react/lib/components/Dropdown';
+
+/**
+* PropertyFieldMultiSelectHost properties interface
+*/
+export interface IPropertyFieldMultiSelectHostProps extends IDropdownProps {
+}