+++ src/propertyFields/multiSelect/index.ts
@@ -1,0 +1,4 @@
+export * from './IPropertyFieldMultiSelect';
+export * from './PropertyFieldMultiSelect';
+export * from './IPropertyFieldMultiSelectHost';
+export * from './PropertyFieldMultiSelectHost';