+++ src/webparts/propertyControlsTest/IPropertyControlsTestWebPartProps.ts
@@ -3,7 +3,8 @@ import { ICheckedTerms } from '../../PropertyFieldTermPicker';
 import { IDateTimeFieldValue } from '../../PropertyFieldDateTimePicker';
 
 export interface IPropertyControlsTestWebPartProps {
-
+  numberValue: number;
+  multiSelect: string[];
   people: IPropertyFieldGroupOrPerson[];
   singleList: string | string[];
   multiList: string | string[];