+++ src/webparts/propertyControlsTest/components/IPropertyControlsTestProps.ts
@@ -6,6 +6,8 @@ import { IPropertyFieldGroupOrPerson } from '../../../PropertyFieldPeoplePicker'
 export interface IPropertyControlsTestProps {
 
   context: WebPartContext;
+  numberValue: number;
+  multiSelect: string[];
   people: IPropertyFieldGroupOrPerson[];
   list: string | string[];
   multiList: string[];