+++ docs/documentation/docs/about/release-notes.md
@@ -1,5 +1,17 @@
 # Releases
 
+## 1.4.0
+
+**New controls**
+
+- `PropertyFieldMultiSelect` got added
+- `PropertyFieldNumber` got added
+
+**Enhancements**
+
+- `PropertyFieldTermPicker` new introduced property to specify to show or hide the term store name: `hideTermStoreName`.
+- `PropertyFieldTermPicker` updated process when terms were unselected based on ID instead of path.
+
 ## 1.3.0
 
 **New controls**